Refactor self reset Occupancy and Motion clusters.
Implement common self reset Occupancy and Motion custom clusters.
MotionWithReset class -- self reset motion cluster, optionally send
event on device occupancy_bus.
MotionOnEvent class -- self reset motion cluster, based on event
received on device motion_bus.
OccupancyOnEvent class -- self reset occupancy cluster, based on event
received on device occupancy_bus.
OccupancyWithReset class -- self reset occupancy cluster, sends an event
to device motion_bus.
