+++ zhaquirks/hivehome/__init__.py
@@ -1,33 +1,11 @@
 """Hive Home."""
-import asyncio
 
-from zigpy.quirks import CustomCluster
-from zigpy.zcl.clusters.security import IasZone
-
-from ..const import CLUSTER_COMMAND, OFF, ZONE_STATE
+from .. import MotionWithReset
 
 HIVEHOME = "HiveHome.com"
 
 
-class MotionCluster(CustomCluster, IasZone):
+class MotionCluster(MotionWithReset):
     """Motion cluster."""
 
-    cluster_id = IasZone.cluster_id
-
-    def __init__(self, *args, **kwargs):
-        """Init."""
-        super().__init__(*args, **kwargs)
-        self._timer_handle = None
-
-    def handle_cluster_request(self, tsn, command_id, args):
-        """Handle the cluster command."""
-        if command_id == 0:
-            if self._timer_handle:
-                self._timer_handle.cancel()
-            loop = asyncio.get_event_loop()
-            self._timer_handle = loop.call_later(30, self._turn_off)
-
-    def _turn_off(self):
-        self._timer_handle = None
-        self.listener_event(CLUSTER_COMMAND, 999, 0, [0, 0, 0, 0])
-        self._update_attribute(ZONE_STATE, OFF)
+    reset_s: int = 30
